Fix CI and refactor commands to use template-based file generation
- Fix CI: Add `.tool-versions` with compatible Flutter 3.41.0.
- Command Refactoring: Introduced `TemplateHelper` and centralized `Templates` to handle file generation across `branding`, `prepare`, and `fastlane` commands.
- `branding`: Fixed incremental resource syncing and modern splash arguments.
- `prepare`: Fixed template compilation by adding missing imports.
- `fastlane`: Modernized build process and fixed artifact paths.
- Process Execution: Switched to real-time log streaming for all tasks.
