+++ lib/commands/prepare_command.dart
@@ -1,4 +1,6 @@
 import 'dart:io';
+import 'package:finvoras_gen/src/templates/templates.dart';
+import 'package:finvoras_gen/src/utils/template_helper.dart';
 import 'package:yaml/yaml.dart';
 import 'package:finvoras_gen/src/core/flutter_generator.dart';
 import 'base_command.dart';
@@ -105,45 +107,7 @@ Future<void> configureDependencies() async => getIt.init();
     await projectService.createDirectories(['lib/core/config']);
     final file = File('lib/core/config/prepare.dart');
 
-    final content = '''
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:flutter_native_splash/flutter_native_splash.dart';
-// TODO: Import your services and DI here
-// import '../../src/di/injection.dart';
-
-Future<void> prepareApp(WidgetsBinding binding) async {
-  // 1. Error widget builder
-  ErrorWidget.builder = (_) => const SizedBox.shrink();
-
-  // 2. Preserve splash screen
-  FlutterNativeSplash.preserve(widgetsBinding: binding);
-
-  // 3. System UI Mode
-  await SystemChrome.setEnabledSystemUIMode(
-    SystemUiMode.manual,
-    overlays: [SystemUiOverlay.bottom, SystemUiOverlay.top],
-  );
-
-  // 4. Dependency Injection
-  // await configureDependencies();
-  // registerAppOrchestratorDependencies();
-  
-  // 5. Core Services Initialization
-  // await AppPathService.instance.init();
-  // await AppKeyStorage.instance.init();
-  // await AppActions.instance.init();
-
-  // 6. Preferred Orientations
-  await SystemChrome.setPreferredOrientations([
-    DeviceOrientation.portraitUp,
-    DeviceOrientation.portraitDown,
-  ]);
-
-  // 7. Remove splash screen
-  FlutterNativeSplash.remove();
-}
-''';
+    final content = TemplateHelper.render(Templates.prepareDart, {});
 
     await file.writeAsString(content);
     print('📝 Created lib/core/config/prepare.dart');
@@ -153,80 +117,10 @@ Future<void> prepareApp(WidgetsBinding binding) async {
     final mainFile = File('lib/main.dart');
     final stack = argResults?['stack'] as String? ?? 'bloc';
 
-    String materialApp;
-    if (stack == 'bloc') {
-      materialApp = '''
-      child: MaterialApp.router(
-        title: 'Flutter Demo',
-        theme: ThemeData(
-          colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
-          useMaterial3: true,
-        ),
-        routerConfig: _router,
-      ),''';
-    } else {
-      materialApp = '''
-      child: GetMaterialApp(
-        title: 'Flutter Demo',
-        theme: ThemeData(
-          colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
-          useMaterial3: true,
-        ),
-        home: const Scaffold(
-          body: Center(child: Text('App Prepared with GetX!')),
-        ),
-      ),''';
-    }
-
-    final content = '''
-import 'package:flutter/material.dart';
-${stack == 'bloc' ? "import 'package:go_router/go_router.dart';\nimport 'package:flutter_bloc/flutter_bloc.dart';" : "import 'package:get/get.dart';"}
-import 'core/config/prepare.dart';
-
-void main() async {
-  final binding = WidgetsFlutterBinding.ensureInitialized();
-  
-  await prepareApp(binding);
-  
-  runApp(const MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return AppOrchestrator(
-$materialApp
-    );
-  }
-}
-
-${stack == 'bloc' ? """
-final _router = GoRouter(
-  routes: [
-    GoRoute(
-      path: '/',
-      builder: (context, state) => const Scaffold(
-        body: Center(child: Text('App Prepared with Bloc & GoRouter!')),
-      ),
-    ),
-  ],
-);
-""" : ""}
-
-class AppOrchestrator extends StatelessWidget {
-  final Widget child;
-  const AppOrchestrator({super.key, required this.child});
-
-  @override
-  Widget build(BuildContext context) {
-    // AppOrchestrator wraps the main app to provide global providers or configurations
-    // For example, MultiBlocProvider for Bloc or initial bindings for GetX
-    return child;
-  }
-}
-''';
+    final content = TemplateHelper.render(Templates.mainDart, {
+      'isBloc': stack == 'bloc',
+      'isGetX': stack == 'getx',
+    });
 
     await mainFile.writeAsString(content);
     print('📝 Updated lib/main.dart (stack: $stack)');
